shorter run for formpost_multipart
P0d now holds PID so we don't have to try get it from OS after receiving signal.

appendOSSStats becomes it's own function so we can call it outside of goroutine for winddown if necessary.

introduced a done channel for DoReqAtmpt (what was I thinking not doing this earlier?) that shuts down all threads.

introduced an array for OSStats that tracks network conns and can be used in future for open files and CPU